requirements.txt
python-dotenv==1.0.0
fastapi==0.100.0
msal==1.22.0
nest-asyncio==1.5.6
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==1.26.16
uvicorn==0.22.0
starlette==0.27.0
pydantic==1.10.11
pydantic_core==2.3.0
mangum==0.17.0